requirements.txt
aiogram
SQLAlchemy
psycopg2
python-dotenv